Fixed up the names of the DE-SynPUF files and the claim types.
Previously, everything had been labeled "Part A Inpatient", "Part A
Outpatient", or "Part B". This was inaccurate. The outpatient claims are
actually under Part B, but for various reasons are stored in the CCW in
the same tables as the Part A inpatient data. The corrected naming
reflects this more complex reality.

CBBD-18