Refactored SampleDataLoader so that it no longer directly uses the DB.
Instead, the new SampleDataPersister can be used to persist things to
the DB, where needed.

This provides **significant** heap memory savings, and also
**significant** speed savings. Previously, it had been impossible to get
the full SAMPLE_1 to load without blowing the heap or segfaulting. Even
with `-Xmx24g`. This now successfully loads all of the SAMPLE_1 data
with "only" an 8 GB heap.

CBBD-45 #done #time 7.1h