First-pass at adding in RIF support: only benes, not many fields.
This will need a lot of rejiggering, I think. For one, I'm not sure that
my current usage of RxJava is providing anything that Java 8's Streams
can't, so I may switch things over to that. For another, this is all
wildly shor on test coverage (only one half-implemented test case right
now), so I highly doubt any of it really works.

That all said... there's a lot of progress here, and it feels like it's
in a solid place to checkpoint.