Add support for mutual SSL auth via Jetty.
Doing this required switching from the cargo-maven2-plugin to the
jetty-maven-plugin, as the latter allows for completely replacing the
configuration that Jetty runs with (and Cargo doesn't). This was needed
in order to ensure that the test configuration mirrors the production
configuration as much as possible.

Note that mutual auth is only supported, not used in the ITs. That will
come in a later commit.

CBBF-30