Refactored benchmarks to stage resources in S3 and support SAMPLE_C.
This ensures that the system running the benchmark isn't uploading all
of the resources over and over for each iteration -- it now just uploads
once for all iterations. This change was necessary as the benchmark was
getting OOM errors trying to upload multiple copies of the SAMPLE_C data
at once. The new approach also incurs less S3 transfer costs, which is a
win.

CBBD-154