Ensure all AWS resources are provisioned in the same AZ.
Without this, AWS was assigning things to random AZs, and I'm pretty sure that this was responsible for a lot of the performance variability that I was observing with the benchmark runs. When I did a benchmark run with this and some other changes (related to IOPS), variability dropped to almost nothing.

CBBD-154