Fixed the tests that relied on the previous S3 key format.
Also fixed the DataSetMonitorWorker such that it properly handles skipping invalid and already-completed data sets, given the new S3 key format. As part of this, I refactored out a new DataSetManifestId class to represent the manifests' identity and processing sort order. That made things a bit cleaner and easier to follow.

CBBD-246