Removed FhirLoaderIT.loadRifDataSampleC() as it's way too slow.
In fact, it'd never even run against an in-memory HSQL DB; you'd have to
use a on-disk one. And even then, it'd take hours or days to complete
and probably need more than a couple terabytes of drive space.

If you really want to run things against SAMPLE_C, use the benchmarks:
they make use of EC2 systems that can handle the large amount of data.

CBBD-256