Made the single-INSERTs load mechanism idempotent.
Note that, without PKs, this slows the load down significantly: a benchmark run I had going for it was only running at 364 records/second vs. 27416 records/second without the idempotency. If PKs are present, a slowdown is still present, though much less severe: it's down to around 5346 records/second.

This slowdown points to the need for making this a configurable option that is only enabled when needed. Haven't done that work yet.

CBBF-53