Log record counts after every RIF file is processed.
This was needed to debug a problem where no entities were actually being saved to the DB. I like the extra assurance it provides, so we'll leave it in. A benchmark run just now indicates that the counting can take two minutes even with PKs enabled, so it's not free, though.

CBBF-53