Refactored to properly model each RifFile in a RifFilesEvent as a nested RifFileEvent.
This simplifies all of the many places in the code that we were passing around a RifFilesEvent and a RifFile pair. It was specifically driven by the desire to have a separate MetricRegistry that tracks the processing of each individual RifFile, rather than tracking things at the data set level.

CBBF-53