Switched from Liquibase to Flyway for schema management.
Flyway isn't quite as automagical, which has pros and cons. But it _does_ have support for running SQL scripts outside of a migration. That support is ugly, but it allows us to run separate "drop/create constraint" scripts, as needed.

My apologies for how ugly this code is if you ever see it; it's a late-night attempt to get something that works. And it does! But I hope to go clean it up soon.

CBBF-53