Very messy refactoring of codeset constants.
The comments at the top of TransformerConstants capture what I'm trying
to accomplish with these changes. This commit is just the first of
several, though, as I still need to review all of this and do some more
cleanup (e.g. sorting TransformerConstants by name).

Also still need to document the codesets being used, which was the
original point of all this.

CBBF-92