Quoted/bracketed fields that might have spaces.
Previous format was impossible to correctly parse, due to lack of clear
field boundaries. We could still be in trouble if we get values with
quotes in them, but that's a low-probability thing with a very high cost
to fix (Wildfly doesn't provide any great facilities for handling it, so
it'd be a lot of extension code).

CBBF-103