Simplify the build versioning on Jenkins.
Trying to make things simple for the "normal" case, which are builds of
the `master` branch. The problem with the previous scheme is that it
breaks dependency versioning across projects. As an example, the
`bluebutton-data-server` POM couldn't depend on the latest `-SNAPSHOT`
build of the `bluebutton-data-model` project, since every Jenkins build
of that project would have had a custom version (as opposed to just
being a normal Maven `-SNAPSHOT`).

On the flip side, the default Maven versioning scheme also won't work
well. Consider the following scenario:

1. A change to the `master` branch is committed and built, resulting in
   a `1.0.0-SNAPSHOT` artifact landing in the repository.
2. A change to the `awesome-new-feature` branch is committed and built,
   resulting in the `1.0.0-SNAPSHOT` version now "pointing" to that
   build, even though it hasn't been merged to `master`. What if this
   change is backwards incompatible?

Depending on the point in time they're made, requests to the artifact
repository to get "the latest build" will bounce around between
branches, almost certainly leading to spurious build failures.

This change attempts to find a middle ground, where `master` branch
builds use the default Maven `-SNAPSHOT` versioning scheme and thus work
as expected. The cost, of course, is that non-`master` branch builds
will be more complex. Specifically, if you are delivering changes that
depend on a non-`master` branch build of a library, you'll have to
update your POM's dependency version number to the exact one that
Jenkins selected during the build.

CBBI-288