Fixed the HSQL config used in tests.
The git commit entry on the previous version makes me think that I
oopsed: I'd meant to set it this was in that earlier commit.

Added some extra documentation explaining the config, because I'm tired
of having to look it up every time I see it and wonder why I set it that
way.

Note: This isn't really related to CBBI-288: I just happened to notice
the discrepancy while working on it.

CBBI-288