Completely reworked Wildfly/JBoss config.
Goal here is to fix intermittent errors that we're seeing out on the
Jenkins server. Specifically, Wildfly was going boom while running the
`:reload` command after configuration. No idea why, but it seems to be a
race condition in Wildfly itself, as it was intermittent and
timing-sensitive.

Instead of using the CLI to configure things, we now just write out a
complete `standalone.xml` file for it in one go. This won't really work
in production, I think, but is fine for local dev/test, where we don't
have to worry about overwriting someone else's config work.

CBBF-131