Reworked the legit EOB.benefitBalance entries.
So far, these are the only two `benefitBalance`s that seem legit. The
other entries look to me more like "claim header adjudication totals",
which I think we ought to map as EOB extensions rather than
`benefitBalance`s. Not addressing that in this commit, though. This just
fixes the two that appear to be mapped appropriately.

CBBF-97