Reduce log noise from codebook lookup failures.
This is a bit hacky, but is the best solution I can think of at the
moment: only log each failure 1x per CcwCodeBookVariable, per app
launch. It'll still leave dupes across servers and restarts, but that
may actually be a good thing, as it may prompt us to actually care about
the lookup failures. We'll see if it ends up being too annoying, I
guess.

CBBF-162