Remove HICN hash from returned EOBs.
It's still present and used behind the scenes for beneficiary mapping,
but it will no longer be sent to applications as part of our EOB
responses.

Why? Because having it in there has always made me a bit uncomfortable.
I'm not convinced that it's a problem, but I'm also unconvinced that
it's safe for us to expose.

BLUEBUTTON-239