Added secondary app servers to play.
Renamed and reorganizaed things quite a bit as part of this, mostly just
for consistency and maintainability.

Note that the secondary test app server is currently (incorrectly)
configured to use the primary test DB server. This is a temporary
workaround for BLUEBUTTON-243.

Also note that I haven't tested these changes yet, so I won't be
surprised if fixes are needed.

BLUEBUTTON-236