Refactored bene history handling a bit.
Just pulled a bit of logic out to a separate method. While investigating
something else, I thought I might need to reuse this logic. Turns out I
don't, but it's a reasonable refactoring anyways, so I'll leave the
change in.

This whole class is a giant mess and needs a more thorough refactoring,
but not today...

BLUEBUTTON-260