Fix exec of server-start.sh, which was hanging.
Apparently, running Bash in "login" mode (per the -l flag) causes it to
wait for all child processes to close their file handles. Specifically,
I think it was waiting for Wildfly/standalone.sh to exit -- which won't
happen.

This fixes the hanging that I was observing locally, but should be
tested on Cygwin, to see if the -l flag was providing any benefit there.

BLUEBUTTON-246