Make idempotency a configurable option.
We use the non-idempotent mode for initial loads. Previously, I'd just
been using custom builds where the `LoadFeature` for it was hardcoded
the way we needed it. That's not a workable approach long-term.

Eventually, it'd be good if idempotent vs. non-idempotent mode could be
automagically configured at runtime by looking at whether the data set
has already been loaded or not, but that's not possible right now.

BLUEBUTTON-467