Restructured things to always hit every stage.
The conditional logic I had before that excluded some stages entirely
was confusing the Jenkins UI. It seems that anytime a build has a
different set of stages, the UI "resets" and won't display the
results-by-stage history for a job.

To solve this, the pipeline now (mostly) no-ops the stages it's going to
skip, but does so from **within** those stages. A likely downside here
is that it'll throw off Jenkins' estimates of how long each stage should
take, but I can't do anything about that.

BLUEBUTTON-469