Refactored and fixed some ugly hidden bugs.
What were the bugs?

1. Tests were brittle because only a certain number of data sets would
   be pre-downloaded and the tests relied on that behavior.
    * When I changed the pre-download count from 3 to 2 in an earlier
      commit, I broke some tests.
2. We were queueing more downloads than we should have, given our disk
   space checks.
3. The `DataSetQueue.cleanup()` deletion of downloaded files wasn't
   safe if we want graceful shutdowns. You can't delete files that might
   be needed by the data set currently being processed.

In general, though, the real problem with the earlier code is that
`DataSetQueue.updatePendingDataSets()` had a bunch of unexpected side
effects that weren't handled well by the rest of the code.

Now, the pre-downloading behavior is a bit more explicit and kicked off
by `DataSetMonitorWorker`, which feels more correct. And the "manage the
queue of pre-downloads" behavior is handled by `S3TaskManager`, which
also feels like the right place for it.

BLUEBUTTON-467