Updated out-of-date SQL constraint scripts.
These are only intended for manual use during an initial load: you might
want to drop the constraints during such a load to speed it up, and then
re-enable the constraints after to ensure correctness of that load and
future data refreshes.

The scripts had gone stale, though, because they're not regularly used.

BLUEBUTTON-467