BLUEBUTTON-266: Paging additions
Changes made to ExplanationOfBenefitResourceProvider.findByPatient to
now return a bundle containing the resources matched by the
beneificiaryId. The bundle is created from the resources pulled from a
new EobBundleProvider class, returning a small sublist of the resources
for each page. Links are added to the bundle for the previous and next
pages as appropriate.