Split DB servers into primaries and secondaries.
The primary DB servers are read-write and should only be used by the
Data Pipeline. The secondary DB servers are asynchronous clones of the
primary and should only be used by the Data Server.

Note: To save money, this split is only_really_ happening in PROD; lower
environments combine the AZ A primary and secondary DB server.

BLUEBUTTON-442