Fix the EC2 instance types and related config.
In particular, I had the wrong instance type for the TEST Data Server
systems, which was causing intermittent crashes.

In general, though, I'm trying to switch things up here to be more
automagical: we just specify in the environment variable files what the
EC2 instances have _total_, and then math our way to the correct Java
configs.

It might take a bit to tweak the functions to be ideal, but should be
less fiddly after that.

BLUEBUTTON-670