BLUEBUTTON-637: Refactoring to use custom runner `Parameterized`
IT now contains two tests: one to generate the "golden" files (tagged
with @Ignore), the other to compare the new endpoint response files to
the approved. Each test uses a list of parameters containing the
endpointId and endpointOperation pairs, performing the test on the
endpointOperation and naming it with the endpointId.