Manage ELBs with Ansible (first draft).
Use Ansible to manage our ELB. The actual goal here is to temporarily
pull one of the test Data Servers out of the ELB's instances, as that
instance is broken. But this will also fix our access logging config,
which isn't standardized across environments.

More importantly, though, this represents a first step towards managing
our AWS infrastructure with code again, instead of by hand. A tiny step,
but an important one to prove out the concept. (That's why IPs and CIDRs
are here -- they're not used yet but likely will be eventually.)

Still a WIP:

* Haven't run it except in `--check` mode yet (and the ELB module
  doesn't support that so just gets skipped).
* Only supports the test environment so far.
* Doesn't have a mechanism for selecting which environments to run
  against; if there _was_ more than one supported, it'd run against all
  of them.

But, that aside, I'm reasonably confident that this should work and want
to get it out for early review before I try running it in the test
environment.

BLUEBUTTON-787