Add/update official docs on our sample data.
The "random" data was not formally submitted to CMS' Data Governance
Board, as it is not significantly derived from actual PII/PHI.
Nonetheless, all updates to that data (and its documentation) must be
reviewed and approved by OEDA leadership. All updates to the
documentation/specification should be made in-place in the
`bluebutton-data-model-rif-samples/dev/bluebutton-ccw-rif-extract-sample-data-random-specification.docx`
file.

The "synthetic" data was formally submitted to CMS' Data Governance
Board, as detailed in the following documents:

*
`bluebutton-data-model-rif-samples/dev/bluebutton-ccw-rif-extract-sample-data-synthetic-dgb-briefing.docx`
*
`bluebutton-data-model-rif-samples/dev/bluebutton-ccw-rif-extract-sample-data-synthetic-dgb-submission.docx`

Those documents should **not** be changed in place. Instead, add a
sibling document/documents to the originals, as needed. All changes to
the "synthetic" data must be reviewed and approved by OEDA leadership
at a minimum, and possibly also CMS' Data Governance Board (at the
discretion of OEDA leadership).

BLUEBUTTON-722