BLUEBUTTON-631: Updating imports and golden files
- The IT and JsonInterceptor both needed import updates after merging
them in for the HAPI-FHIR version update
- The metadata.json golden file has been updated to the correct URLs, a
change that was fixed in HAPI-FHIR release 3.0 and was what spurred us
to want to make this change, so it is expected.
- The remaining golden files have changed only because I had to merge
this branch with master, which brought in Dave's changes on PR#126. You
can see the golden files now have NPI display values where appropriate.