BLUEBUTTON-802: Replace ignored fields with filler text
First step, changes to the IT for a place to get a handle on the JSON
fields that will need to be updated. This in turn slightly changes the
format of all golden files. Committing them now so that in subsequent
commits we can look specifically at the fields that are being replaced.