Mostly complete: access log parsing & analysis.
Just a set of scripts for reading in and parsing access log data,
importing it into SQLite, and some basic analysis views for it there.

Useful in the context of moving to structured logs, because the AWK
parsing script could easily be used to convert our access logs to JSON
is we need to do that.

Mostly though, I created all this as a way to analyze the performance of
our various query types out in PROD.

BLUEBUTTON-945