Log client SSL principals in access.log.
Main change here is the addition of the `<security-domain
name="bluebutton-data-server" />` configuration entry, and the
`jboss-web.xml` file that activates/selects it for this app. Most
everything else just supports that switch.

How does this result in client SSL principals being logged? It makes the
"`%r`" variable in our access log pattern (in `standalone.xml`) actually
work. With the security domain configured this way, to read the client
certificate, the remote use is available and captured in our access
logs.

Note the companion PR for `ansible-role-bluebutton-data-server` that has
already been merged and deployed. It adds the security domain to our AWS
environmets' configs.

https://github.com/CMSgov/ansible-role-bluebutton-data-server/pull/7

BLUEBUTTON-945