Pull Data Servers from use before updating.
This prevents the Data Servers from receiving queries (that they return
errors for) while they're being updated.

Note: this is a simplification of a similar branch that I was working on
with Dave: I found a couple modules (`ec2_metadata_facts` and
`elb_instance`) that work a lot easier and better.

BLUEBUTTON-984