Restart JBoss, rather than calling :reload.
I think the issue causing this intermittent bug is the use of `:reload`
inside an `if ... end-if` block. But that's just a guess, so the safest
thing to do is always restart the whole service, rather than reload.

I also decided to go from three config blocks to just two, which should
make the extra delay from a full restart slightly less painful.

BLUEBUTTON-1057