Use :reload instead of service restart.
My previous commits/PRs just didn't pan out: restarting the service
right after config changes led to _different_ race condition errors with
JBoss.

We'll see, but I suspect this method here is the way to do it: use
`:reload`, but ensure it's the last thing in the script; it can't even
be inside a block at the end of the script.

But I'm fully prepared to find out I'm wrong here, too. JBoss is just
full of surprises like that.

BLUEBUTTON-1057