Make bene HICN and MBI fields lazy and unordered.
Making them lazily loaded is a bit of a security improvement: we don't
want to have the data unless/until it's actually needed.

Making them unordered is 1) required in order to avoid Hibernate
throwing `MultipleBagFetchException` when trying to join/load the (lazy)
fields, and 2) probably a more honest modeling of the data.

BLUEBUTTON-865