Fixed EOBs-by-HICN with naive but working DB query.
We're back to three separate DB queries again, which rather stinks. But,
on the bright side, it works!

I have an idea for how to cut it down to two queries that I may try to
commit on this branch next.

Note: the commented-out code was an attempt to use native SQL queries to
go with a much better single-query approach. Unfortunately, I was unable
to get this to work with JBoss/Hibernate.

BLUEBUTTON-865
BLUEBUTTON-1110