Add NDJSON access and DB query logs.
The primary change here is the addition of a
`bluebutton-server-app-log-access.json` NDJSON log file, which will have
exactly one entry for every response sent out. This log makes heavy use
of the MDC to include the following in those entries:

* All the information in our current access logs.
* All request and response HTTP headers.
* Details on all DB queries run as part of the response and how long
  they took.
* Details on all JPA operations run as part of the response, how long
  they took, and how many records they returned.

In addition, a `bluebutton-server-app-log-database.json` log was added,
which, in the default configuration, will log all slow or unexpected
database queries, with their full query text.

BLUEBUTTON-1177