Corrected binary names for new structure.
Tried to keep this to just documentation and the bits of code that would
otherwise have been broken by all the renames. Hard to know if I got
everything, though, given how easy it is for the old names to hide in
amongst the package, class, variable names, etc.

BLUEBUTTON-1086