Make roles, plays, and Jenkinsfile compatible.
This is more of a draft at the moment, as there are a ton of moving
parts here and I haven't tested any of it together yet.

Nonetheless, the goal here is to get the Ansible roles and plays
compatible with each other and also the new Jenkinsfile build.

BLUEBUTTON-1086