Configure autoformatting of Java sources.
As part of `mvn initialize` (and later phases) in Java builds, we run
the autoformatter.

As part of Git's pre-commit hook, we verify that the formatting is
correct and bail out of the commit if it's not.

Also updated Eclipse config in devenv docs to configure Eclipse to
(mostly) apply the same styling, via its builtin autoformatter.

BLUEBUTTON-1086