Fixed New Relic to work with Jetty.
Maybe, anyways: this hasn't been tested yet.

Looks like New Relic wouldn't have been working earlier, with JBoss.
This fixes all that.

This is out of scope for the rest of the "switch to Jetty" changes, but
I didn't realize that until I was too far in to untangle things. So.

BLUEBUTTON-1112