Fix deploy error caused by ELB health checks.
I'm not 100% sure what's causing it, but HealthAPT deploys were failing
saying that the instance wasn't healthy. I think it's because we were
stopping the service first, before switching from JBoss to Jetty, and
then not giving the ELB time to notice that it was healthy again. This
should fix that.

BLUEBUTTON-1112