Create BFD Pipeline EC2 instance after SGs.
When I rolled out my previous PR to the `test` environment, the BFD
Pipeline app failed to launch the first time, because the RDS security
group it depended on wasn't in place.

This is a rather gross hack, but should fix it. It's be much simpler if
Terraform just supported `depends_on` for modules, but it apparently
doesn't yet.

BLUEBUTTON-1288