Fixed ownership of S3TaskManager.
Previously, it'd been owned and managed by each DataSetMonitorWorker,
which is pretty odd: we don't want separate async and download queues
for each scanning run; those should be more global resources.

Now, the S3TaskManager is owned by the DataSetMonitor, which is a
singleton service. That feels safer and more appropriate.

I don't _think_ the previous behavior was causing any thread leaks, but
that's only because we happened to be setting the thread pools to time
out idle threads. That's a pretty subtle thing, and I'd rather not rely
on it.

BLUEBUTTON-1361