Fixed bug in S3 downloads causing test errors.
The errors were intermittent, but frequent. I still can't quite believe
this is true, but it turns out that the **only** thread-safe operation
on a `Future` is `Future.get()`. The race condition we were running into
was causing the task to be marked as cancelled even after it had
completed, which the `TaskExecutor` was noticing and complaining about.

BLUEBUTTON-1361