Integrated Plaid into Java BFD Server.
The BFD Server now forwards all PDEs-by-bene searches to the Plaid
application.

This is only a basic integration, at least so far:
* It's going to be slow as the BFD Server is de-serialising all of the
  results from Plaid and then re-serialising them.
* It only covers PDE EOB searches, not reads.
* Plaid isn't done, and so this integration effectively breaks the BFD
  Server.

But that breakage is actually what I wanted to achieve here: we can now
use the BFD Server's integration tests to verify whether or not Plaid is
working as expected. This saves us from having to go re-create all of
those tests in Rust right now.