Enable JVM option required for flamegraphs.
Per https://medium.com/netflix-techblog/java-in-flames-e763b3d32166,
this is needed to generate CPU flamegraphs.

Also note this quote about the option: "We’ve found in tests that this
costs between 0 and 3% extra CPU, depending on the workload."

That seems an acceptable cost to me, given the usefullness of flame
graphs for performance profiling.

BLUEBUTTON-1644