Return synthetic resources with negative IDs
A regex for parsing the format of EoB and Coverage resource
read requests did not successfully parse synthetic
Resource read requests due to their negative IDs.
The result was that no synthetic EoB or Coverage resources
were found on Read requests.

This updates those regex to allow for the IDs of these
Resources to be negative.