Adding Pipfiles.
They're useful for running Ansible, etc. locally, and I've had them
laying around locally forever. Eventually, it'd be good to try and
consolidate them and also remove the older requirements.txt files
that are still there, but... baby steps.

(Note that this has nothing to do with the referenced JIRA ticket,
except that I was working on it when I finally decided to clean this
up.)

BLUEBUTTON-1644