BLUEBUTTON-1841 Create indexes for partdcontactnumber columns (#224)
Create indexes for partdcontactnumber columns

A standard index was created for each column using the `concurrently` feature of postgres in order to limit it's affects on reads and writes during the creation of the index.

Searching for beneficiaries based on partDContractNumber<month>Id proved slow over large data sets. This is a simple solution that would naturally go away when coverage data is stored in a more robust way.

A hash index was considered but ultimately not used due to a caution listed here https://www.postgresql.org/docs/9.6/indexes-types.html

A partial index was considered but ultimately not used due to added complexity without an apparent improvement in performance.