BLUEBUTTON-2002 Return synthetic resources with negative IDs (#258)
* Return synthetic resources with negative IDs

A regex for parsing the format of EoB and Coverage resource
read requests did not successfully parse synthetic
Resource read requests due to their negative IDs.
The result was that no synthetic EoB or Coverage resources
were found on Read requests.

This updates those regex to allow for the IDs of these
Resources to be negative.

* Throw IllegalArgumentException on invalid eobId

This changes the exception from ResourceNotFoundException
to IllegalArgumentException when the eobIdText does not
match the expected Regex pattern. This parameter is
invalid rather than indicating an eobId was not found.

* Throw IllegalArgumentException for invalid ID

This changes the exception from ResourceNotFoundException
to IllegalArgumentException when the coverageIdText does not
match the expected Regex pattern. This parameter is
invalid rather than indicating an covargeId was not found.

* Update IT tests for negative and invalid IDs

This updates the IT tests for Coverage and EOB read()
methods to test that the ID parameters work as expected
for negative IDs and IDs that do not pass the regex
pattern for a valid format.

* Update comments per Karl's suggestion

Co-authored-by: David Gage <whytheplatypus@gmail.com>