Add (hilariously) basic test coverage for Synthea.
Adds support for generating data via a special branch/build of Synthea.
For now, this data is just run through `RifLoaderIT` to verify it works.

This is a hilariously rough first draft, with a number of issues and
TODOs noted in the changes. Notably, building Synthea is **slow**: takes
15-30 minutes on my laptop.

BFD-458