Adds max_instance_lifetime to fhir asg
This change will automatically recycle instances after the amount of seconds
(7 days initially) set in the 'max_instance_lifetime' variable.

You can read more about this feature here https://docs.aws.amazon.com/autoscaling/ec2/userguide/asg-max-instance-lifetime.html
, but in short, AWS should (smartly) recycle instances (one at a time)
so that no instance will live longer than defined.

Depending on the number of running instances, the ASG will start recycling instances in
advance so that all instances are recycled before reaching this ttl.

**warning**
There is a possibility that it could attempt to recycle all instances at
once if all instances are over this age (7 days currently). So, care
should be taken the first time this is implemented. I.e., do a deploy
first so we have new instances, merge, and deploy again to implement.

Also, keep in mind that running instance will be terminated no matter
if a user is logged into it or not. E.g., if you want to work on
something in test, detach the instance from the ASG first so your work
will not dissapear into the ether.

**why?**

First and formost, to prevent stale instances from causing issues with partners as
we have seen in the past.

Secondly, recycling instances is a good practice that helps to identify
problems early on (e.g., chaos theory) and can help provide a better
security posture.