Aide fixup (#463)
* Aide fixup

There is a cron entry baked into the gold image that runs `aide` and attempts to send the results over email. This is problematic because:
1. Email is not a secure method of sending sensitive data.
2. The results of the aide check are too large so the email fails.
3. The aide database needs to be updated after installing/updating packages.

This patch removes the failing cron job and updates the aide database so our ops can run it on live instances as needed.

* refactor shell command to be backwards compatible with older tf

* Update main.yml

Also for cron command.