Adds posix users/groups to ETL box
When partners mount our EFT EFS file systems they are rooted to a
directory and all file operations are mapped to a high numbered posix
user and group. This simply adds these users and groups (starting with
BCDA). The user and groups are system users with no ability to login.