Updates Ansible versions (#533)
Ansible versions are defined throughout the BFD project in a variety of
places. Some of which are likely unused or only used locally by single
engineers. However, since these versions frequently come up in our synk
or dependabot scans, this work simply updates all the versions I could
find where it made sense to update.

Note: Because pip is broken on FIPS enabled RHEL7 systems, we are
currently using the ansible available in the default yum packages, which
is 2.9.18. This is the version used on our build servers in prod, so it
makes sense to try to pin the same versions.

I did not update a few "requirements.freeze.txt" files as those should
be artifacts from running 'pip freeze' (ideally we would not be using
this anywhere). But, if they come up in a security scan, we can either
remove/add to gitignore, or update the versions.