Try to fix intermittent bfd-server socket errors.
This one is a bit harder to debug, just because it's less frequent, but
I think I may have tracked down the issue. I think it's similar to the
other `SocketException` errors that were just fixed in `bfd-pipeline`:
the server was binding to a different host/interface than we were trying
to connect to in our tests.

BFD-704