Fixed ConcurrentModficationException bug.
`PipelineJobRecordStore.jobRecords` can be modified by almost every
thread in the application, so was definitely at risk for
`ConcurrentModificationException`s. It's honestly surprising that I
didn't hit any in the tests. Fortunately, though, none of the use cases
for it care that it's only eventually consistent (from the perspective
of any one thread), so the fix is just a simple type change.

Kudos to @brianburton for pointing this one out!