Resolve intermittent SocketExceptions in tests. (#540)
We weren't being consistent in how we checked for and reserved the port,
across everything. As a result, we were often reserving ports that were
available on 127.0.0.1, but not on "*", which is what HSQL was actually
binding to.

This may have been particularly rough on CMS laptops, as one of the
applications installed for CMS Mac OS users likes to bind a whole bunch
of high-number random ports. Things seemed to get worse recently, where
I could almost never get a build to succeed due to these
`SocketException`s.

BFD-704

Co-authored-by: Chris Brune <64486927+cbrunefearless@users.noreply.github.com>