BFD-704: Made BFD Pipeline orchestration more flexible (#539)
Made BFD Pipeline orchestration more flexible.

Everything is now explicitly part of a "job", and scheduled as such.
This will make it much simpler for new jobs to be added, e.g. those
related to DC Geo and Synthea.

Note: removed `RifLoaderIdleTasks` and friends as it turns out that code
was quite broken. It apparently only worked accidentally; it was not
working as intended, at all. Since we no longer need those jobs, it
wasn't worth trying to refactor and fix all of it.

BFD-704