Refactored and fixed DB query monitoring.
Previously, the measured elapsed time for queries included the time
taken to acquire a connection from the pool, which ended up
**dominating** the reported times, causing them all to be bad
overestimates. There's still some unavoidable async-scheduling
overestimation in this refactored version, but it's now much smaller:
tens or hundreds of milliseconds, as opposed to the earlier hours.

In addition, the code is now quite a bit simpler and less redundant,
which is quite nice.

BFD-1071