DCGEO-117: Initial service date range implementation (#697)
Add ability to filter search results by a specified service date range. For FISS claims we are filtering on the stmtCovToDate value and for MCS claims we are filtering on the idrHdrToDate. We are trying to maintain backwards compatibility with service-date filtering on ExplanationOfBenefit records, with the exception that we are moving this logic to filter on the database rather than in Java code.