BFD-1566: Implement CCW RIF Filtering for 2022 Enrollment Issues (#980)
We need a filtering mechanism in our loads such some beneficiary records are temporarily skipped: only those with an `ENRLMNT_RFNC_YR` of "2022" or `NULL` will be processed (which, during the calendar year of 2022, is most of them).

As part of this filtering, we are implementing an assumption that no non-2022 `INSERT` beneficiary records will be received, as skipping those would also require skipping their associated claims, which is additional complexity that we want to avoid. If any  such records are encountered, the load will go boom. This filtering is an inelegant hack to work around upstream data issues, and will hopefully only be in place temporarily. See the code that uses this field in `RifLoader` for details. This filtering is being made configurable so as to not invalidate all of our existing test coverage.

Co-authored-by: Logan Mitchell <logan.mitchell@adhocteam.us>
Co-authored-by: Chris Trotter <chris.trotter@adhocteam.us>