Properly specify requirements using pip-compile to generate secure requirements.txt from top-level dependencies; add psycogreen as top-level dependency
"pip-compile" is a tool included with the "pip-tools" package and is recommended for use with managing Python requirements. With it, we can get very similar behavior to something like NPM's "package.json" and "package-lock.json". In our case, the "requirements.in" is similar to package.json and "requirements.txt" is similar to package-lock.json. We also get the benefit of using pip's secure mode, allowing us to specify exact hashes for dependencies such that we can guarantee that downloaded dependencies are exactly the version we expect. This also guards against increasingly common package hijacking.