Remove comparison metadata for all response time percentiles under 90%
The response times under 90% usually either vary wildly or are so small (1 - 10 ms) as to be meaningless when comparing against a baseline. Frequently, warn thresholds for 50% or 66% would be exceeded for response times in the 1 - 10 ms range, meaning that a deployment could fail for a task having a 50% of 2 - 3 ms instead of 1 ms which would be a serious false negative.