BFD-1773: Cleanup unused or deprecated Terraform related to metrics and alarms and update EoB SLO alarms (#1470)
* Remove unnecessary, deprecated metric filters that filter upon access.txt -- these have been superseded by access.json metric filters

* Remove unnecessary bfd_server_alarm module; remove old SLO alarms superseded by work done in BFD-1772; remove deprecated metric filters

* Remove unused cw_metric_alarms module

Alarms defined within the module are not referenced anywhere in the BFD source nor are they applied to any environment's set of CloudWatch alarms. In addition, they have not been touched for over 3 years.

* Remove unused cw_metric_filters module defining metric filters on the per-env access.txt log group(s)

This module is not referenced anywhere else in the BFD project's Terraform nor are any of the metric filters defined in the module applied to the access.txt log group(s). Additionally, this module has not been touched for over 3 years.

* Remove unused bfd_dashboard module

There are no references to this module anywhere else in the BFD project's Terraform; additionally, the dashboard(s) that this module creates are not currently applied to our AWS infrastructure

* Remove unused ec2_alarms module

No references to this module in BFD's other Terraform nor are any of the alarms applied to our infrastructure

* Remove unused nat_alarms module

No references to this module in BFD's other Terraform nor are any of the alarms applied to our infrastructure

* Remove unused rds_alarms module

No references to this module in BFD's other Terraform nor are any of the alarms applied to our infrastructure

* Add new "eob with resources" metric filters and metrics; update EoB SLO alarms to use new "with resources" metrics

* Update CloudWatch SLO dashboard to use EoB with resources metrics